Refactors into LWRP
Instead of using a recipe/attributes, we now have an LWRP that is a bit
more flexible. It allows you to specify a global owner of watched
directories, and override this per directory.

We also allow a custom syncid that falls back to the hostname. This
fixes an issue on hostnames that don't conform to csync2's syncid
limitations ([a-z0-9]+).

In addition, more thorough testing has been added, and debian-7.9 has
been added to the testing matrix.

Resolves #1
Resolves #2